Release version v0.223.0
  - [391659496] [KMS] Add custom retry for spanner clients.
  - [0][MIC] fixed PrivateKeyClient log
  - [384621359] [MIC] Add Terraform script to turn up an Open...
  - [0] [CMRTIO] Remove slf4jsimple from CPIO deps
  - [387301152][KMS] Switching Load Balancer 5xx alerts back ...
  - [380085779] [CMRTIO] Introduce MetricType in CustomMetric...
  - [375263642][MIC] Add new google_cloud_run_service_iam_mem...
  - [380085779] [INFRA] Rename enable_native_metric_aggreatio...
  - [382325173][KMS] create spanner backup schedule. Requirem...
  - [389926816][MIC] Fix SpannerKeyDb db reading to properly ...
  (And 72 more changes)

GitOrigin-RevId: 475062aabf1288edd3b03fcf26933333242a4786

+++ cc/core/nosql_database_provider/src/gcp/gcp_spanner_utils.h
@@ -152,7 +152,8 @@ class GcpSpannerUtils {
     } else if (const float* float_val = std::get_if<float>(&attribute_value)) {
       // Technically floats are not supported by spanner::Value but it will be
       // converted to a double.
-      out_value = google::cloud::spanner::Value(*float_val);
+      out_value =
+          google::cloud::spanner::Value(static_cast<double>(*float_val));
     } else if (const double* double_val =
                    std::get_if<double>(&attribute_value)) {
       out_value = google::cloud::spanner::Value(*double_val);
